data types, numbe

+++ Variables.java
@@ -1,6 +1,6 @@
 public class Variables {
     public static void main(String[] args) {
-        double myNumber = 123;
+        int myNumber = 123;
         String myString = "kayak";
         System.out.println(myNumber + " " + myString);
       }
